requirements.txt
transformers
pytorch-crf
torch
joblib
sentencepiece
scikit-learn